Organized and removed unnecessary files
Organized repetitive files and fixed references.
