+++ EmergencyDataCollectionApp/src/components/CustomForms/CustomDateTimePickerComponent/CustomDateTimePickerComponent.js
@@ -2,7 +2,7 @@ import DateTimePicker from "@react-native-community/datetimepicker";
 import React, { useState } from "react";
 import { View, TouchableOpacity, Text } from "react-native";
 
-import { formatDate } from "../../../screens/MYNReportPage/components/formatDate";
+import { formatDate } from "./formatDate";
 import Theme from "../../../utils/Theme";
 
 const CustomDateTimePickerComponent = ({
