+++ EmergencyDataCollectionApp/src/components/CustomForms/NativeBase/CustomTextArea/HelperText.js
