--- EmergencyDataCollectionApp/src/screens/MYNReportPage/components/formatDate.js
@@ -1,12 +1,0 @@
-export const formatDate = (date) => {
-  if (!date) {
-    return "No date selected";
-  }
-  return `${(date.getMonth() + 1).toString().padStart(2, "0")}/${date
-    .getDate()
-    .toString()
-    .padStart(2, "0")}/${date.getFullYear()} ${date
-    .getHours()
-    .toString()
-    .padStart(2, "0")}:${date.getMinutes().toString().padStart(2, "0")}`;
-};
